backend/requirements/base.txt
# Base
pytz==2020.1
python-slugify==4.0.1
Pillow==7.2.0
psycopg2==2.7.4 --no-binary psycopg2

# Django
django==3.0

# Django REST framework
djangorestframework==3.11.1

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==5.2.0

# Celery
redis>=3.5.3
django-redis==4.12.1
celery==4.4.7
flower==0.9.5

# jwt
pyjwt==1.7.1

# CORS
django-cors-headers==3.5.0
